adding sample apps
