+++ samples/HelloEDAM/res/values/arrays.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="esdk__main_list">
+        <item>@string/esdk__image_picker</item>
+        <item>@string/esdk__simple_note</item>
+    </string-array>
+</resources>
