+++ samples/HelloEDAM/res/values/strings.xml
@@ -6,12 +6,25 @@
     <string name="label_app_title">Evernote EDAM Android Demo</string>
     <string name="label_select_button">Select Image</string>
     <string name="label_message">No image selected</string>
-    <string name="label_save_button">Save to Evernote</string>
+    <string name="label_save_button">Save</string>
     <string name="err_image_selected">Select Image</string>
     <string name="err_retrieving_resource">Error retrieving resource</string>
     <string name="msg_retrieved_resource">Retrieved resource from server</string>
-    <string name="label_log_in">Sign in to Evernote</string>
-    <string name="label_log_out">Log out of Evernote</string>
-    <string name="loading">Loading...</string>
+    <string name="esdk__loading">Loading...</string>
     <string name="esdk__pick_notebook">Pick a notebook</string>
+    <string name="label_login">Login</string>
+    <string name="label_logout">Sign out</string>
+    <string name="esdk__image_picker">Image Picker</string>
+    <string name="esdk__simple_note">Simple Note</string>
+    <string name="label_select_notebook_button">Select Notebook</string>
+    <string name="title">Note Title</string>
+    <string name="content">Note Content</string>
+    <string name="empty_content_error">Note title and content cannot be empty</string>
+    <string name="note_saved">Note saved to Evernote</string>
+    <string name="ok">Ok</string>
+    <string name="simple_note">Simplenote</string>
+    <string name="image_picker">Image Picker</string>
+    <string name="error_listing_notebooks">Error listing notebooks</string>
+    <string name="error_creating_notestore">Error creating notestore</string>
+    <string name="error_saving_note">Error saving note</string>
 </resources>
